requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
colorama==0.4.6
Django==5.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
sqlparse==0.5.1
tzdata==2024.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
